# Fix performance regression in DataServiceContext DefaultResolveType method