@gathogojr 2022-12-08
> The code looks good to me. I would request some clarifications on the performance measurements:
> 
> * Are the figures shared measuring the expected/best case (i.e., the type is in the `targetAssembly`)?
> * If the provided figures are for the expected case, could you share figures for the worst case if possible (the type cannot be resolved)
> * Could you run the tests on a version of the client before the changes that introduced the regression were made, just so we can have an idea how close it is to initial expectations, if not the same (or better), at least for the common case (the type does exist in the target or entry assembly).

@habbes Failure to find a type to use for materialization when deserializing a payload results into an exception. However, I was able to create a simple benchmark to measure the time the `DefaultResolveType` method takes if a type is not found. The `DefaultResolveType` method is protected that's why I did it this way.
```csharp
public class DataServiceContextBenchmark : DataServiceContext
{
    [Benchmark]
    [IterationCount(targetIterationCount: 100)]
    public void BenchmarkResolveType()
    {
        for (var i = 0; i < 250; i++)
        {
            this.DefaultResolveType("NonExistentType", "NS.Models", "NS.Models");
        }
    }
}
```

Stats:

**Before this fix:**
| Run |               Method |     Mean |     Error |   StdDev |
|-----|--------------------- |---------:|----------:|---------:|
| 1 | BenchmarkResolveType | 4.559 us | 0.3813 us | 1.124 us |
| 2 | BenchmarkResolveType | 2.599 us | 0.0847 us | 0.2319 us |
| 3 | BenchmarkResolveType | 5.856 us | 0.7412 us | 2.174 us |
| 4 | BenchmarkResolveType | 4.339 us | 0.1353 us || 0.3882 us |
| 5 | BenchmarkResolveType | 3.541 us | 0.2066 us | 0.5828 us |

**After this fix:**
| Run |               Method |     Mean |     Error |    StdDev |
|-----|--------------------- |---------:|----------:|----------:|
| 1 | BenchmarkResolveType | 2.912 us | 0.0930 us | 0.2653 us |
| 2 | BenchmarkResolveType | 3.260 us | 0.1559 us | 0.4572 us |
| 3 | BenchmarkResolveType | 2.262 us | 0.0394 us | 0.1136 us |
| 4 | BenchmarkResolveType | 4.269 us | 0.3640 us | 1.068 us |
| 5 | BenchmarkResolveType | 2.603 us | 0.1108 us | 0.3160 us |

The results are a bit inconsistent across runs but it'd be right to say their is an observable improvement after the fix